Build-Fix: Code fuer ESP-IDF 6.1 / GCC 15 lauffaehig gemacht
- main.c: Doppelte Funktionen entfernt (init_hardware, init_nvs, initialize_sntp)
- main.c: esp_sleep API aktualisiert (ext1-Wakeup fuer Deep Sleep, gpio_wakeup_enable fuer Light Sleep)
- main.c: SNTP auf esp_sntp_* umgestellt (deprecated APIs)
- main.c: configure_deep_sleep_wakeup in Nacht-Deep-Sleep eingebunden (Wakeup-Bug behoben)
- watchdog: watchdog_stop()/watchdog_start() hinzugefuegt (TWDT ab-/anmelden vor/nach Sleep)
- pir.h/servo.h: EVENT_MAX_COUNT Makros in Header verschoben (von web_server.c genutzt)
- pir.c: ungenutzte motion_start_time entfernt
- web_server.c: duration_ms Cast fuer %u (GCC 15 Format-Check)
- heap_monitor.c: esp_get_total_heap_size -> heap_caps_get_total_size (IDF 6.0 API)
- error_log.h: ERR_HEAP_LOW Fehlercode ergaenzt
- wifi.c: lwip/ip4_addr.h Include fuer IP4_ADDR
- wifi: wifi_deinit -> wifi_module_deinit (Namenskonflikt mit esp_wifi-Library)

+++ components/main/web_server.c
@@ -528,7 +528,7 @@ static esp_err_t pir_events_handler(httpd_req_t *req)
             (i > 0) ? "," : "",
             (unsigned long long)pir_events_buffer[i].timestamp_ms,
             time_str,
-            pir_events_buffer[i].duration_ms
+            (unsigned int)pir_events_buffer[i].duration_ms
         );
     }
 
@@ -574,7 +574,7 @@ static esp_err_t servo_events_handler(httpd_req_t *req)
             (i > 0) ? "," : "",
             (unsigned long long)servo_events_buffer[i].timestamp_ms,
             time_str,
-            servo_events_buffer[i].duration_ms
+            (unsigned int)servo_events_buffer[i].duration_ms
         );
     }
